+++ libs/designer/src/lib/ui/configuretemplate/tabs/profileTab.tsx
@@ -18,20 +18,35 @@ export const profileTab = (
   disabled,
   content: <TemplateManifestForm />,
   footerContent: {
-    primaryButtonText: resources.PreviousButtonText,
-    primaryButtonOnClick: () => {
-      dispatch(selectWizardTab(constants.CONFIGURE_TEMPLATE_WIZARD_TAB_NAMES.PARAMETERS));
-    },
-    secondaryButtonText: resources.NextButtonText,
-    secondaryButtonOnClick: () => {
-      dispatch(selectWizardTab(constants.CONFIGURE_TEMPLATE_WIZARD_TAB_NAMES.REVIEW));
-    },
-    thirdButtonText: intl.formatMessage({
-      defaultMessage: 'Save',
-      id: 'GGmFte',
-      description: 'The description for saving the profile tab content to the service provider',
-    }),
-    thirdButtonOnClick: onSave,
-    thirdButtonDisabled: disableSave,
+    buttonContents: [
+      {
+        type: 'button',
+        text: resources.PreviousButtonText,
+        onClick: () => {
+          dispatch(selectWizardTab(constants.CONFIGURE_TEMPLATE_WIZARD_TAB_NAMES.PARAMETERS));
+        },
+      },
+      {
+        type: 'button',
+        text: resources.NextButtonText,
+        onClick: () => {
+          dispatch(selectWizardTab(constants.CONFIGURE_TEMPLATE_WIZARD_TAB_NAMES.REVIEW));
+        },
+      },
+      {
+        type: 'divider',
+      },
+      {
+        type: 'button',
+        text: intl.formatMessage({
+          defaultMessage: 'Save',
+          id: 'GGmFte',
+          description: 'The description for saving the profile tab content to the service provider',
+        }),
+        appreance: 'primary',
+        onClick: onSave,
+        disabled: disableSave,
+      },
+    ],
   },
 });