+++ libs/designer/src/lib/ui/panel/templatePanel/quickViewPanel/tabs/workflowTab.tsx
@@ -29,7 +29,7 @@ export const workflowTab = (
   workflowId: string,
   clearDetailsOnClose: boolean,
   onPrimaryButtonClick: (() => void) | undefined,
-  { templateId, workflowAppName, isMultiWorkflow }: Template.TemplateContext,
+  { templateId, workflowAppName, isMultiWorkflow, showCreate, showCloseButton }: Template.TemplateContext,
   onClose?: () => void
 ): TemplateTabProps => ({
   id: constants.TEMPLATE_PANEL_TAB_NAMES.WORKFLOW_VIEW,
@@ -41,32 +41,43 @@ export const workflowTab = (
   tabStatusIcon: undefined,
   content: <WorkflowPanel workflowId={workflowId} />,
   footerContent: {
-    primaryButtonText: intl.formatMessage({
-      defaultMessage: 'Use this template',
-      id: '5szzYP',
-      description: 'Button text to create workflow from this template',
-    }),
-    primaryButtonOnClick: () => {
-      LoggerService().log({
-        level: LogEntryLevel.Trace,
-        area: 'Templates.overviewTab',
-        message: 'Template create button clicked',
-        args: [templateId, workflowAppName, `isMultiWorkflowTemplate:${isMultiWorkflow}`],
-      });
-      dispatch(openPanelView({ panelView: TemplatePanelView.CreateWorkflow }));
-      onPrimaryButtonClick?.();
-    },
-    secondaryButtonText: intl.formatMessage({
-      defaultMessage: 'Close',
-      id: 'FTrMxN',
-      description: 'Button text for closing the panel',
-    }),
-    secondaryButtonOnClick: () => {
-      dispatch(closePanel());
-      if (clearDetailsOnClose) {
-        dispatch(clearTemplateDetails());
-      }
-      onClose?.();
-    },
+    buttonContents: [
+      {
+        type: 'button',
+        text: intl.formatMessage({
+          defaultMessage: 'Use this template',
+          id: '5szzYP',
+          description: 'Button text to create workflow from this template',
+        }),
+        appreance: 'primary',
+        onClick: () => {
+          LoggerService().log({
+            level: LogEntryLevel.Trace,
+            area: 'Templates.overviewTab',
+            message: 'Template create button clicked',
+            args: [templateId, workflowAppName, `isMultiWorkflowTemplate:${isMultiWorkflow}`],
+          });
+          dispatch(openPanelView({ panelView: TemplatePanelView.CreateWorkflow }));
+          onPrimaryButtonClick?.();
+        },
+        hide: !showCreate,
+      },
+      {
+        type: 'button',
+        text: intl.formatMessage({
+          defaultMessage: 'Close',
+          id: 'FTrMxN',
+          description: 'Button text for closing the panel',
+        }),
+        onClick: () => {
+          dispatch(closePanel());
+          if (clearDetailsOnClose) {
+            dispatch(clearTemplateDetails());
+          }
+          onClose?.();
+        },
+        disabled: !showCloseButton,
+      },
+    ],
   },
 });